(UNTESTED) Implement tree printing functions
While trying to write avlToString, I noticed that I couldn't figure out
a way to incorporate the traversal method, and that's because we don't
need it. The avlToString function returns a string representing the
structure of the given tree, and does not visit the tree's nodes in a
particularly special order.